Refactor pipeline pre-setup and run context handling
This commit introduces significant refactoring to improve the handling of pipeline pre-setup and run contexts. Key changes include:

- Asynchronous methods for obtaining pipeline information and pre-setup, enhancing responsiveness.
- Removal of nullable types for certain parameters to improve type safety.
- New properties in the `IRunContext` interface for detailed run context information.
- Simplified context-checking logic by accessing properties directly.
- Restructuring of context classes (`BumpContext`, `CommitContext`, etc.) for better clarity.
- Added error handling for unset environment variables to ensure graceful failures.
- Improved overall code organization with better separation of concerns.
