+++ NukeBuildHelpers/Pipelines/Local/LocalPipeline.cs
@@ -1,8 +1,11 @@
-﻿using NukeBuildHelpers.Common.Enums;
+﻿using Nuke.Common.IO;
+using NukeBuildHelpers.Common;
+using NukeBuildHelpers.Common.Enums;
 using NukeBuildHelpers.Entry.Interfaces;
 using NukeBuildHelpers.Entry.Models;
 using NukeBuildHelpers.Pipelines.Common.Interfaces;
 using NukeBuildHelpers.Pipelines.Common.Models;
+using System.Text.Json;
 
 namespace NukeBuildHelpers.Pipelines.Local;
 
@@ -12,19 +15,23 @@ internal class LocalPipeline(BaseNukeBuildHelpers nukeBuild) : IPipeline
 
     public BaseNukeBuildHelpers NukeBuild { get; set; } = nukeBuild;
 
-    public PipelineInfo GetPipelineInfo()
+    public Task<PipelineInfo> GetPipelineInfo()
     {
-        return new()
+        return Task.FromResult(new PipelineInfo()
         {
             Branch = NukeBuild.Repository.Branch,
             TriggerType = TriggerType.Local,
-            PullRequestNumber = 0,
-        };
+            PullRequestNumber = null,
+        });
     }
 
-    public PipelinePreSetup? GetPipelinePreSetup()
+    public async Task<PipelinePreSetup> GetPipelinePreSetup()
     {
-        return null;
+        await NukeBuild.RunPipelinePreSetup();
+        var pipelinePreSetup = JsonSerializer.Deserialize<PipelinePreSetup?>(
+        (Nuke.Common.NukeBuild.TemporaryDirectory / "NUKE_PRE_SETUP.json")
+            .ReadAllText()) ?? throw new Exception("NUKE_PRE_SETUP is empty");
+        return pipelinePreSetup;
     }
 
     public Task PreparePreSetup(AllEntry allEntry)
@@ -32,27 +39,29 @@ internal class LocalPipeline(BaseNukeBuildHelpers nukeBuild) : IPipeline
         return Task.CompletedTask;
     }
 
-    public Task FinalizePreSetup(AllEntry allEntry, PipelinePreSetup? pipelinePreSetup)
+    public Task FinalizePreSetup(AllEntry allEntry, PipelinePreSetup pipelinePreSetup)
+    {
+        (Nuke.Common.NukeBuild.TemporaryDirectory / "NUKE_PRE_SETUP.json")
+            .WriteAllText(JsonSerializer.Serialize(pipelinePreSetup, JsonExtension.SnakeCaseNamingOption));
+        return Task.CompletedTask;
+    }
+
+    public Task PreparePostSetup(AllEntry allEntry, PipelinePreSetup pipelinePreSetup)
     {
         return Task.CompletedTask;
     }
 
-    public Task PreparePostSetup(AllEntry allEntry, PipelinePreSetup? pipelinePreSetup)
+    public Task FinalizePostSetup(AllEntry allEntry, PipelinePreSetup pipelinePreSetup)
     {
         return Task.CompletedTask;
     }
 
-    public Task FinalizePostSetup(AllEntry allEntry, PipelinePreSetup? pipelinePreSetup)
+    public Task PrepareEntryRun(AllEntry allEntry, PipelinePreSetup pipelinePreSetup, Dictionary<string, IRunEntryDefinition> entriesToRunMap)
     {
         return Task.CompletedTask;
     }
 
-    public Task PrepareEntryRun(AllEntry allEntry, PipelinePreSetup? pipelinePreSetup, Dictionary<string, IRunEntryDefinition> entriesToRunMap)
-    {
-        return Task.CompletedTask;
-    }
-
-    public Task FinalizeEntryRun(AllEntry allEntry, PipelinePreSetup? pipelinePreSetup, Dictionary<string, IRunEntryDefinition> entriesToRunMap)
+    public Task FinalizeEntryRun(AllEntry allEntry, PipelinePreSetup pipelinePreSetup, Dictionary<string, IRunEntryDefinition> entriesToRunMap)
     {
         return Task.CompletedTask;
     }
