--- NukeBuildHelpers/RunContext/Interfaces/ICommitContext.cs
@@ -1,8 +1,0 @@
-﻿namespace NukeBuildHelpers.RunContext.Interfaces;
-
-/// <summary>
-/// Represents a context that includes information specific to commit runs.
-/// </summary>
-public interface ICommitContext : IPipelineContext
-{
-}
