--- NukeBuildHelpers/RunContext/Interfaces/ILocalContext.cs
@@ -1,8 +1,0 @@
-﻿namespace NukeBuildHelpers.RunContext.Interfaces;
-
-/// <summary>
-/// Represents a context that includes information specific to local runs.
-/// </summary>
-public interface ILocalContext : IRunContext
-{
-}
