--- NukeBuildHelpers/RunContext/Models/CommitContext.cs
@@ -1,7 +1,0 @@
-﻿using NukeBuildHelpers.RunContext.Interfaces;
-
-namespace NukeBuildHelpers.RunContext.Models;
-
-internal class CommitContext : PipelineContext, ICommitContext
-{
-}
