--- NukeBuildHelpers/RunContext/Models/LocalContext.cs
@@ -1,7 +1,0 @@
-﻿using NukeBuildHelpers.RunContext.Interfaces;
-
-namespace NukeBuildHelpers.RunContext.Models;
-
-internal class LocalContext : RunContext, ILocalContext
-{
-}
