--- NukeBuildHelpers/RunContext/Models/PullRequestContext.cs
@@ -1,9 +1,0 @@
-﻿using NukeBuildHelpers.Entry.Models;
-using NukeBuildHelpers.RunContext.Interfaces;
-
-namespace NukeBuildHelpers.RunContext.Models;
-
-internal class PullRequestContext : VersionedContext<PullRequestReleaseVersion>, IPullRequestContext
-{
-    PullRequestReleaseVersion IPullRequestContext.AppVersion { get => AppVersion; }
-}
